+++ node/express-router.ts
@@ -25,7 +25,10 @@ import {
 export class ExpressRouter {
 	router: FetchRouter;
 	constructor(options: NodeRouterOptions = {}) {
-		this.router = new FetchRouter({ routes: options.routes ?? [] });
+		this.router = new FetchRouter({
+			routes: options.routes ?? [],
+			errorHandler: (error, request) => this.onRouteError(error, request),
+		});
 	}
 
 	middleware(): ExpressRequestHandler {
@@ -39,6 +42,10 @@ export class ExpressRouter {
 
 	getResponse(request: Request): Promise<Response> {
 		return this.router.getResponse(request);
+	}
+
+	onRouteError(error: unknown, request: Request): void {
+		console.error("Fatal Route Error", error);
 	}
 
 	respond(res: ExpressResponse, response: Response): void {